+++ .github/release-notes/release-notes.md
@@ -3,6 +3,8 @@ Provides data factories and a test suite base class to inherit from.
 ## 🛠 Enhancements
 
 - [#10](https://github.com/zhofre/test-fabric/issues/10) Added easy way to generate names and emails.
+- [#10](https://github.com/zhofre/test-fabric/issues/10) Added easy way to generate company names.
+- [#10](https://github.com/zhofre/test-fabric/issues/10) Added easy way to generate countries and cities.
 
 ## 🐞 Bug Fixes